move dockicon code to separate source file
